refactor: nested layout 네비게이션 외부주입으로 변경
