+++ src/components/index.ts
@@ -1,10 +1,10 @@
 export { default as Input } from './common/Input';
-export { default as Layout } from './common/Layout';
 export { default as LikeButton } from './common/LikeButton';
 export { default as LoadingSpinner } from './common/LoadingSpinner';
 export { default as Symbol } from './common/Symbol';
 export { default as Textarea } from './common/Textarea';
 export { default as TitleLogo } from './common/TitleLogo';
+export { default as Layout } from './common/layout/Layout';
 
 export { default as ProfileImage } from './images/ProfileImage';
 export { default as TweetImage } from './images/TweetImage';
